# Conflicts:
#	src/llm_wiki/rocksdb_wiki_store.cpp

+++ src/acceleration/compute_backend.cpp
@@ -93,19 +93,19 @@ PartialBatchResult IVectorBackend::batchKnnSearchSafe(
 
     // Build a contiguous buffer of valid query vectors
     std::vector<float> validQueries(validIndices.size() * dim);
-    for (size_t i = 0; i < validIndices.size(); ++i) {
+    for (size_t i = 0; i <static_cast<int>(validIndices.size()); ++i) {
         const float* src = queries + validIndices[i] * dim;
         std::copy(src, src + dim, validQueries.data() + i * dim);
     }
 
     // Dispatch valid queries through the deterministic batchKnnSearch
     auto batchResults = batchKnnSearch(
-        validQueries.data(), validIndices.size(), dim,
+        validQueries.data(),static_cast<int>(validIndices.size()), dim,
         vectors, numVectors, k, useL2
     );
 
     // Map results back to the original query indices
-    for (size_t i = 0; i < validIndices.size(); ++i) {
+    for (size_t i = 0; i <static_cast<int>(validIndices.size()); ++i) {
         KnnQueryResult& qr = result.queryResults[validIndices[i]];
         qr.neighbors = std::move(batchResults[i]);
         qr.status    = AccelerationErrorCode::Success;
