+++ src/sharding/converged_slo_metrics.cpp
@@ -68,8 +68,8 @@ double ConvergedSLOMetrics::getLatencyP99(const std::string& op_type) const {
     std::sort(sorted.begin(), sorted.end());
     
     size_t index = static_cast<size_t>(sorted.size() * 0.99);
-    if (index >= sorted.size()) {
-        index = sorted.size() - 1;
+    if (index >= static_cast<int>(sorted.size())) {
+        index = static_cast<int>(sorted.size()) - 1;
     }
     return sorted[index];
 }
